Rename network module to networks. Move networkmanager helper functions into a new file. Add some network tests.
