--- data/etc/plinth/modules-enabled/network
@@ -1,1 +1,0 @@
-plinth.modules.network
