+++ data/etc/plinth/modules-enabled/networks
@@ -1,0 +1,1 @@
+plinth.modules.networks
