+++ plinth/modules/networks/templates/connections_delete.html
@@ -32,7 +32,7 @@
     <input type="submit" class="btn btn-md btn-primary"
            value="Delete {{ name }}"/>
 
-    <a href="{% url 'network:index' %}" role="button"
+    <a href="{% url 'networks:index' %}" role="button"
        class="btn btn-md btn-primary">Cancel</a>
   </form>
