+++ plinth/modules/networks/templates/connections_edit.html
